Move generateName() into a new class
