+++ Wurst Client/src/tk/wurst_client/gui/alts/GuiAltEdit.java
@@ -22,6 +22,7 @@ import org.lwjgl.opengl.GL11;
 import tk.wurst_client.Client;
 import tk.wurst_client.alts.Alt;
 import tk.wurst_client.alts.AltUtils;
+import tk.wurst_client.alts.NameGenerator;
 import tk.wurst_client.alts.SkinStealer;
 import tk.wurst_client.utils.MiscUtils;
 
@@ -115,7 +116,7 @@ public class GuiAltEdit extends GuiScreen
 				}else
 					errorTimer = 8;
 			}else if(clickedButton.id == 3)
-				nameBox.setText(AltUtils.generateName());
+				nameBox.setText(NameGenerator.generateName());
 			else if(clickedButton.id == 4)
 				displayText = SkinStealer.stealSkin(newAlt.name);
 			else if(clickedButton.id == 5)
