+++ Wurst Client/src/tk/wurst_client/gui/alts/GuiAltLogin.java
@@ -23,6 +23,7 @@ import org.lwjgl.opengl.GL11;
 
 import tk.wurst_client.Client;
 import tk.wurst_client.alts.AltUtils;
+import tk.wurst_client.alts.NameGenerator;
 import tk.wurst_client.alts.SkinStealer;
 import tk.wurst_client.utils.MiscUtils;
 
@@ -103,7 +104,7 @@ public class GuiAltLogin extends GuiScreen
 				else
 					errorTimer = 8;
 			}else if(clickedButton.id == 3)
-				nameBox.setText(AltUtils.generateName());
+				nameBox.setText(NameGenerator.generateName());
 			else if(clickedButton.id == 4)
 				displayText = SkinStealer.stealSkin(nameBox.getText());
 			else if(clickedButton.id == 5)
